requirements.txt
dm_haiku==0.0.1a0
jax==0.1.59
jaxlib==0.1.39
dm_env==1.2
pytest
git+https://github.com/deepmind/bsuite.git
git+https://github.com/deepmind/rlax
mock==3.0.5
coverage==5.0.1
pytest-cov==2.8.1
grpcio==1.26.0
grpcio-tools==1.26.0
protobuf==3.9.2
